JEL-226-rework: Load messages configuration for the Transport factory in the extension
